requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
openai==1.17.0
python-dotenv==1.0.1
